refactor: separate note text
