+++ src/components/PopupNote/PopupNote.scss
@@ -52,20 +52,6 @@
     }
   }
 
-  &__text {
-    font: inherit;
-
-    color: $isabelline;
-    resize: none;
-    background: none;
-    border: none;
-    outline: none;
-
-    &::placeholder {
-      color: $darker-quick-silver;
-    }
-  }
-
   &__link {
     position: absolute;
     top: 0;
@@ -75,11 +61,5 @@
     width: 100%;
     height: 100%;
 
-  }
-
-  .hashtag {
-    padding: 2px 5px;
-    
-    background-color: $tag-color;
   }
 }
